requirements.txt
numpy==2.0.0
pandas==2.2.2
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.4.2
scipy==1.14.0
setuptools==69.5.1
torch==2.3.0
tqdm==4.66.4
xgboost==2.1.0
